[Core] PropertiesFile-Management
+ renamed PropertiesFileReader -> PropertiesFileManager
+ created new class PropertiesFileReader
+ moved reading-responsibility to PropertiesFileReader
+ fixed Javadoc in PropertiesFileInterface and renamed method getPropValues() -> getValue()
+ applied changes to classes that were previously using the PropertiesFileReader
